Update _Main.py
edited events

+++ _Main.py
@@ -17,9 +17,9 @@ async def on_ready():
 async def on_message(msg):
     if msg.channel.id == 534754067998834688 and (
         msg.content.lower() != 'f.verify' or msg.content.lower() != 'ff verify'):
-        return await msg.delete()
-    else:
-        await bot.process_commands(msg)
+        if msg.author.id != 191036924570501120:
+            return await msg.delete()
+    await bot.process_commands(msg)
 
 
 if __name__ == '__main__':
